WARC-218 SeekableDataFileInputStream.getPosition returns long and is now stored internally as a long

+++ gov.usgs.jem.swfmm.grid/src/gov/usgs/jem/swfmm/grid/input/SeekableDataFileInputStream.java
@@ -56,7 +56,7 @@ public interface SeekableDataFileInputStream extends DataInput, Closeable
 	 * @return the position of the pointer within the file
 	 * @since Oct 27, 2016
 	 */
-	int getPosition();
+	long getPosition();
 
 	/**
 	 * Create a new {@link SeekableDataFileInputStream} using the provided file
